requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytesseract>=0.3.8
Pillow>=8.3.1
numpy>=1.21.2
pdf2image>=1.16.0
reportlab>=3.6.1
python-dotenv>=0.19.0
aiofiles>=0.7.0
httpx>=0.19.0
opencv-python>=4.5.3
googletrans==4.0.0-rc1 